fix: AdminOnly Methods for editing the items in a workspace work again
