+++ src/main/java/com/example/roommate/values/domainValues/RoomNumber.java
@@ -1,4 +1,8 @@
 package com.example.roommate.values.domainValues;
 
 public record RoomNumber(String number) {
+    @Override
+    public String toString() {
+        return "RoomNumber" + number;
+    }
 }
